requirements.txt
Django==3.2.15
psycopg2==2.9.3
djangorestframework==3.13.1
markdown==3.4.1
django-filter==22.1
requests
